Update README.md
Welcome to "Advanced Data Analysis," a comprehensive and demanding course designed to bridge the gap between the rich, dynamic world of social sciences and the transformative power of computing technologies. This course is meticulously structured to introduce you to a range of computational tools and programming fundamentals, specifically tailored for applications in social science research. Starting with the basics of setting up essential software like Anaconda Navigator and Google Colab, the course progresses through the nuances of Python programming, data manipulation using Pandas, and dives into the realms of statistical analysis, and machine learning. Each week, we'll explore critical themes, from web scraping and spatial analysis to cutting-edge topics in machine learning like Artificial Neural Networks. The course is enriched with a carefully curated selection of readings, providing both theoretical backdrops and practical insights into each topic. By the end of this journey, you'll be equipped with the skills and knowledge to harness computational techniques in social science research, opening doors to innovative methodologies and insightful analyses in your future endeavors.